Split pointcloud_utilities from wytham-initial
The code in the pointcloud_utilities package was developed and
stored in the same repository as the various notebooks (and other
associated files, data etc) that it was used with, making it
confusing to deal with. This package has now been split from said
'wytham-initial' directory, so that it can be maintained
seperately, publically and with less clutter.